RET-2574: Rename case assignment to local office service and move it to service package
